Refactor config classes with getName for cam config and doubles instead of floats

+++ src/main/java/dinwiddy/ipcam/framegrab/config/ICaptionServerConfig.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface ICaptionServerConfig {
 
-	public float getDefaultFrequencySeconds();
+	public double getDefaultFrequencySeconds();
 	
 	public String getDefaultCaptionText();
